URL slug came from the index of the rendered class list instead of being equal to the class Id on props).
On the teacher dashboard, made sure that the courses were being fetched properly based on the currently logged in user’s id.  This was done in server/api/users.js.  Instead of fetching a user object based on req.params.userId and calling .getCourses() on it, I’m just searching (based on req.params.userId) directly through the Course table.