Added some new users for testing purposes, split the course intro and more information strings by \n and mapped over them to create a new div for each line break on the student classboard, filtered out enrolled classes under the  section title on the student dashboard, made the links on the student dashboard redirect to the correct classroom page (because originally the URL slug came from the index of the rendered class list instead of being equal to the class Id on props).
On the teacher dashboard, made sure that the courses were being fetched properly based on the currently logged in user’s id.  This was done in server/api/users.js.  Instead of fetching a user object based on req.params.userId and calling .getCourses() on it, I’m just searching (based on req.params.userId) directly through the Course table.