Fixed a lot of reference errors; make sure all the ids match up!
When creating courses in seed.data.js, we also want to specify the ID.

Also fixed a problem where the teacher classboard wasn’t rendering the assignments for that particular class, and was just getting the index of the classes list which is jumbled up both in the database and perhaps from seed.js.

For the teacher list of courses, need to use something other than the getUserCoursesThunk in client/store/user.js.  It’s not returning the courses taught by that user, but rather is returning the courses in which the user is enrolled.

Added alert message - username/password is incorrect! - for user/login authentication.
Also included toggle lecture feature, toggle attendance features on the Teacher Classboard.  When we toggle a lecture, we show/hide the student dashboard with the course id passed to it as props, which is then used to render the class component on the same page without changing the URL.  Similarly we toggle attendance.

Thunk creator and API route and action to the reducer for fetching all the assignments for a course, by courseId.  Mapping over these assignments from the Redux state, in the teacher assignments view.