No longer mapping over all schedules on the teacher classboard, because it’s supposed to be for a single class.
Instead, we fetch a single course object and use it in the teacherclassboard component as well as pass it down to the student classboard component as props.

Now rendering list of students at the top of the single classboard view

Currently, I’m refactoring things so we that we don’t always get course number/id, etc. from the this.props.location object.  The thing is, this.props.location only exists if we’re rendering the component by accessing its route through the URL; if we’re just rendering it through the teacher classboard, then there isn’t any location attribute and so we need to pass down the course object from teacherclassboard as props in order to access values on it.

So I open a socket on the course ID, and if we’re actually viewing the component through the URL, which is what we do as a student, then we still set firstNameForSocket (which was declared as an empty variable) to be equal to this.props.location.state.firstName.  otherwise, we declare it to be this.props.userInherited.firstName (when we’re accessing it as a component within the teacherclassboard).

So I basically changed the firstName variable (what you see in the chat when you type something) to a dynamic variable instead of checking this.props.state.location, which may or may not exist depending on how we render the component.

Also added static (non-functioning) buttons - Record, MuteAll - In Chat, Mute Specific - In Chat, Remove Message, Cancel Chat for student, Poll (survey) - to the studentClassboard (it shouldn’t even be called studentClassboard, it should be called just classboard) component, and these buttons only render if the person is logged in as a teacher.

The first slash is very important when making an axios.get request in a thunk creator.