There are actually four aspects of attendance - we want to be able to refer to the courseId.
Created Attendance table: currently can post to the Attendance table and retrieve attendance data by courseId.  The attendance status (present, tardy, or absent) is not required to be null because sometimes the teacher may not know for sure whether or not a student is there, and may want to revise their attendance later on.

One direction to improve attendance would be to also add a PUT route in order to modify existing values.