The three aspects of attendance - student id, attendance status, and current date/time - are being mapped to state.  Now we just need to map them to the database.
There are actually four aspects of attendance - we want to be able to refer to the courseId.

Created Attendance table: currently can post to the Attendance table and retrieve attendance data by courseId.  The attendance status (present, tardy, or absent) is not required to be null because sometimes the teacher may not know for sure whether or not a student is there, and may want to revise their attendance later on.

One direction to improve attendance would be to also add a PUT route in order to modify existing values.

Now fetching past attendance history from the database, divided by horizontal line rules; currently I need to fetch the name of the student based on that id, which is on the props for the Attendance component.

Formatted the time returned by Moment.js (for example, Wednesday, August 26th 2020, 3:26:07 pm), and converted the student id returned by the getAttendanceForCourse thunk (which fetches based on an inherited course id from the teacher classboard component) to firstName + ‘ ‘ + lastName.

The show past attendance button/data is now updated without having to refresh the page

Moved the submit attendance function, this.props.getAllAttendanceForThisCourse(this.props.courseIdInherited), outside of the forEach loop which creates the attendance array, to prevent multiple POST requests being made for each student.

The attendance array is now being retrieved from this.state.

On TeacherAssignmentView, creating a drop-down menu to select specific assignments and map the selected value to state on change.

Mapping over the course’s assignment list from the Redux State to create different options on the drop-down menu.

Filtering out the assignments list depending on the selected value located on this.state.

Also added ‘show all’ feature to the assignments drop-down menu.
 Add Total Points (totalPoints) and Percent of Total Grade (percentTotalGrade) to the Assignments model and create some examples in seed.data.js.   Parse the deadline which is already included in the Assignments model and database.   Reflect this on the front-end.

Fetching deadline, total points, and percentage of grade for each assignment from the database.  Also imported moment on TeacherAssignmentView so that we can format the due date into a more readable format.

The Students component on the teacher classboard should have a drop-down menu for all/specific.  For the selected student(s), we should display the points they earned, the total points, calculate the percent their score represents for them, and the grade their score represents based on that percent.

Essentially, the TeacherClassboard component passes data to each component as props.

On the embedded TeacherAssignementByStudentView component, I’ve made two drop-down menus for filtering by student and filtering by assignment.  Both of them have an option for selecting all students or all assignments.